@@ -10,42 +10,51 @@
 		"ngrok": "ngrok http 3000",
 		"start": "node --max-old-space-size=4096 -- node_modules/ts-node/dist/bin -P tsconfig.json app.ts",
 		"start:dev": "ts-node-dev -P tsconfig.json app.ts",
-		"lint": "eslint '*/**' --quiet --fix",
-		"test": "echo \"Error: no test specified\" && exit 1"
+		"build": "tsc",
+		"lint": "eslint '*/**' --quiet",
+		"lint:fix": "npm run lint -- --fix",
+		"test": "echo \"Error: no test specified\" && exit 0",
+		"migrate": "do the thing"
 	},
 	"license": "MIT",
 	"dependencies": {
-		"@slack/bolt": "^4.1.1",
-		"@slack/web-api": "^7.7.0",
+		"@slack/bolt": "^4.2.0",
+		"@slack/web-api": "^7.8.0",
+		"@types/pg": "^8.11.11",
+		"adze": "^2.2.0",
 		"clark": "0.0.6",
+		"convict": "^6.2.4",
 		"cron": "^3.2.1",
-		"dotenv": "^16.4.5",
+		"date-fns": "^4.1.0",
 		"image-charts": "^6.1.19",
-		"jest": "^29.7.0",
-		"lodash": "^4.17.21",
-		"moment": "^2.30.1",
-		"mongoose": "^8.8.3",
-		"newrelic": "^12.8.0",
+		"lodash.map": "^4.6.0",
+		"lodash.samplesize": "^4.2.0",
+		"lodash.take": "^4.1.1",
+		"newrelic": "^12.11.3",
 		"pg": "^8.13.1",
+		"pluralize": "^8.0.0",
 		"reflect-metadata": "^0.2.2",
 		"slack-block-builder": "^2.8.0",
 		"token-buddy": "^1.0.0",
-		"ts-node": "^10.9.2",
-		"typeorm": "^0.3.20",
-		"typescript": "^5.7.2"
+		"ts-node": "^10.9.2"
 	},
 	"devDependencies": {
+		"@eslint/eslintrc": "^3.2.0",
+		"@eslint/js": "^9.19.0",
+		"@slack/types": "^2.14.0",
 		"@types/convict": "^6.1.6",
-		"@types/lodash": "^4.17.13",
-		"@types/node": "^22.10.1",
-		"@typescript-eslint/eslint-plugin": "^8.16.0",
-		"@typescript-eslint/parser": "^8.16.0",
-		"convict": "^6.2.4",
-		"eslint": "^9.16.0",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-prettier": "^5.2.1",
-		"ngrok": "^4.3.3",
-		"prettier": "^3.4.1",
-		"ts-node-dev": "^2.0.0"
+		"@types/lodash": "^4.17.15",
+		"@types/node": "^22.13.0",
+		"@types/pluralize": "^0.0.33",
+		"eslint": "^9.19.0",
+		"eslint-config-prettier": "^10.0.1",
+		"eslint-plugin-jest": "^28.11.0",
+		"eslint-plugin-prettier": "^5.2.3",
+		"jest": "^29.7.0",
+		"ngrok": "^5.0.0-beta.2",
+		"prettier": "3.4.2",
+		"ts-node-dev": "^2.0.0",
+		"typescript": "^5.7.3",
+		"typescript-eslint": "^8.23.0"
 	}
 }
