@@ -1,4 +1,6 @@
 export interface AuditTags {
-  updatedAt?: Date;
-  updatedBy?: string;
+	updatedAt?: Date;
+	updatedBy?: string;
+	createdAt?: Date;
+	createdBy?: string;
 }
