@@ -1,36 +1,10 @@
-import { Schema, Document, Model, Connection } from 'mongoose';
-
-export interface IScoreLog extends Document {
-  from: string;
-  to: string;
-  date: Date;
-  channel: string;
-  scoreChange: number;
-  reason?: string;
+export interface IScoreLog {
+	from: string;
+	to: string;
+	date: Date;
+	channelId: string;
+	channelName?: string;
+	// eslint-disable-next-line @typescript-eslint/no-redundant-type-constituents
+	scoreChange: 1 | -1 | number;
+	reason?: string;
 }
-
-export const ScoreLogSchema = new Schema({
-  from: String,
-  to: String,
-  channel: String,
-  reason: String,
-  date: {
-    type: Date,
-    default: new Date(),
-    index: -1,
-  },
-  scoreChange: Number,
-});
-
-ScoreLogSchema.index({ to: 1, from: 1, date: -1 });
-
-export interface ScoreLogInterface extends IScoreLog {
-  // instance methods
-}
-
-export interface ScoreLogInterfaceModelInterface extends Model<ScoreLogInterface> {
-  // static methods
-}
-
-export const ScoreLog = (conn: Connection) =>
-  conn.model<ScoreLogInterface, ScoreLogInterfaceModelInterface>('scoreLog', ScoreLogSchema);
