@@ -1,17 +1,17 @@
 export const actions = {
-  hometab: {
-    admin_settings: 'HOMETAB_ADMIN_SETTINGS',
-    admin_settings_submit: 'HOMETAB_ADMIN_SETTINGS_SUBMIT',
-    sync_admins: 'HOMETAB_SYNC_ADMIN',
-    user_settings: 'HOMETAB_USER_SETTINGS',
-    user_settings_submit: 'HOMETAB_USER_SETTINGS_SUBMIT',
-  },
-  wallet: {
-    level_up_confirm: 'LEVEL_UP_CONFIRM',
-    level_up_cancel: 'LEVEL_UP_CANCEL',
-  },
-  shortcuts: {
-    message: 'message_plusplus',
-    global: 'global_plusplus',
-  },
+	hometab: {
+		admin_settings: 'HOMETAB_ADMIN_SETTINGS',
+		admin_settings_submit: 'HOMETAB_ADMIN_SETTINGS_SUBMIT',
+		sync_admins: 'HOMETAB_SYNC_ADMIN',
+		user_settings: 'HOMETAB_USER_SETTINGS',
+		user_settings_submit: 'HOMETAB_USER_SETTINGS_SUBMIT',
+	},
+	wallet: {
+		level_up_confirm: 'LEVEL_UP_CONFIRM',
+		level_up_cancel: 'LEVEL_UP_CANCEL',
+	},
+	shortcuts: {
+		message: 'message_plusplus',
+		global: 'global_plusplus',
+	},
 };
