@@ -1,28 +1,28 @@
 export const blocks = {
-  hometab: {
-    admin: {
-      basic: {
-        admins: 'hometab-a-pointdPal-admins',
-        companyName: 'hometab-a-company-name',
-        notificationChannel: 'hometab-a-notif-channel',
-        falsePositiveNotificationChannel: 'hometab-a-false-notif-channel',
-        scoreboardChannel: 'hometab-a-scoreboard-channel',
-        formalPraiseUrl: 'hometab-a-formal-url',
-        formalPraiseMod: 'hometab-a-formal-mod',
-      },
-      crypto: {},
-    },
-    user: {
-      crypto: {
-        walletAddress: 'hometab-u-crypto-wallet',
-      },
-    },
-  },
-  shortcuts: {
-    message: {
-      recipients: 'shortcut-message-recipients',
-      operator: 'shortcuts-message-operator',
-      reason: 'shortcuts-message-reason',
-    },
-  },
+	hometab: {
+		admin: {
+			basic: {
+				admins: 'hometab-a-pointdPal-admins',
+				companyName: 'hometab-a-company-name',
+				notificationChannel: 'hometab-a-notif-channel',
+				falsePositiveNotificationChannel: 'hometab-a-false-notif-channel',
+				scoreboardChannel: 'hometab-a-scoreboard-channel',
+				formalPraiseUrl: 'hometab-a-formal-url',
+				formalPraiseMod: 'hometab-a-formal-mod',
+			},
+			crypto: {},
+		},
+		user: {
+			crypto: {
+				walletAddress: 'hometab-u-crypto-wallet',
+			},
+		},
+	},
+	shortcuts: {
+		message: {
+			recipients: 'shortcut-message-recipients',
+			operator: 'shortcuts-message-operator',
+			reason: 'shortcuts-message-reason',
+		},
+	},
 };
